+++ src/discovery/index.ts
@@ -1,2 +1,2 @@
 export * from './metadata-discovery-provider';
-export * from './swagger-discovery-provider';
+export { SwaggerDiscoveryProvider, type SwaggerDiscoveryOptions } from './swagger-discovery-provider';