+++ src/metadata/index.ts
@@ -1,3 +1,5 @@
 export * from './decorators';
+export * from './decorators-reflect';
+export * from './controller-decorators';
 export * from './registry';
 export * from './helpers';