+++ src/swagger/index.ts
@@ -1,0 +1,1 @@
+export { enrichSpecWithAiMetadata, useAiSwaggerEnricher } from './spec-enricher';